minor styles
