--- src/hooks/debounceHook.jsx
@@ -1,20 +1,0 @@
-import React, { useState, useEffect } from 'react';
-
-const useDebounce = (value, timeout, callback) => {
-    const [timer, setTimer] = useState(null);
-
-    const clearTimer = () => {
-        if (timer) clearTimeout(timer);
-    };
-
-    useEffect(() => {
-        clearTimer();
-
-        if (value && callback) {
-            const newTimer = setTimeout(callback, timeout);
-            setTimer(newTimer);
-        }
-    }, [value]);
-};
-
-export default useDebounce;
